+++ src/app/detail/[id]/detail.tsx
@@ -1,11 +1,13 @@
 import Image from "next/image";
 import detailDummy from "../../../../public/detail_dummy.png";
-export default function Detail() {
+import { ProductDetail } from "./page";
+export default function Detail({ data }: { data: ProductDetail }) {
+  const contentImg = `
+    ${data.content}
+  `;
   return (
     <>
-      <div className="my-5">
-        <Image src={detailDummy} width={428} height={450} alt="막걸리 이미지" />
-      </div>
+      <div className="my-5" dangerouslySetInnerHTML={{ __html: contentImg }}></div>
     </>
   );
 }
